float to total shell

+++ apps/core/models.py
@@ -87,9 +87,8 @@ class venta(models.Model):
     sucursal = models.ForeignKey(Sucursal, db_column='Sucursal_id') # Field name made lowercase.
     fecha = models.DateTimeField( auto_now_add = True, db_column = "fecha" )
 
-
     folio = models.IntegerField( db_column = "folio")
-    total = models.IntegerField( db_column = "total")
+    total = models.FloatField( db_column = "total")
     total_productos = models.IntegerField( db_column = "total_productos")
     url_reporte =   models.CharField(max_length=500L, blank=True , db_column = "url_reporte")
